+++ src/styles/colors.css
@@ -68,7 +68,6 @@
     --light-theme-background-off-white: var(--off-white);
     --light-theme-faint-text: var(--white-two);
     --light-theme-modal-supplemental-text: var(--charcoal-grey);
-    --light-theme-modal-footer-bg: #f6f4ff;
     --light-theme-modal-btn-default-color: var(--dark-four);
     --light-theme-btn-default-disabled-bg: var(--heather);
 }
